# appFastLog/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FastLog API",
        "description": "API para gerenciamento de entregas via API REST",
        "contact": {
            "email": "admin@fastlog.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API FastLog"
        },
        {
            "url": "http://localhost/api/",
            "description": "API Local Server"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza autenticação do usuário e retorna um token de acesso",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@exemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|laravel_sanctum_token_hash"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciais inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/token": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Gera um novo token de acesso para um usuário autenticado",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "novo_token_de_acesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Lista todos os pedidos",
                "operationId": "9368500233e75a120cb606e08d7a06e4",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pedido"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/api/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Retorna um pedido específico",
                "operationId": "69873c1dd29645d07e53a4c4db9057e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Pedido": {
                "required": [
                    "numeroPedido",
                    "destinatarioNome",
                    "destinatarioEndereco",
                    "destinatarioTelefone",
                    "itemDescricao",
                    "entregador_id",
                    "status_pedido_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID do pedido",
                        "type": "integer",
                        "format": "int64"
                    },
                    "numeroPedido": {
                        "description": "Número único do pedido",
                        "type": "string",
                        "maxLength": 50
                    },
                    "destinatarioNome": {
                        "description": "Nome do destinatário",
                        "type": "string",
                        "maxLength": 100
                    },
                    "destinatarioEndereco": {
                        "description": "Endereço de entrega",
                        "type": "string",
                        "maxLength": 200
                    },
                    "destinatarioTelefone": {
                        "description": "Telefone do destinatário",
                        "type": "string"
                    },
                    "itemDescricao": {
                        "description": "Descrição do item a ser entregue",
                        "type": "string",
                        "maxLength": 500
                    },
                    "entregador_id": {
                        "description": "ID do entregador responsável",
                        "type": "integer"
                    },
                    "status_pedido_id": {
                        "description": "ID do status do pedido",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Pedidos",
            "description": "API Endpoints de gerenciamento de pedidos"
        },
        {
            "name": "Autenticação",
            "description": "Autenticação"
        }
    ]
}